+++ Web/Scotty.hs
@@ -25,14 +25,14 @@ module Web.Scotty
     , capture, regex, function, literal
       -- ** Accessing the Request and its fields
     , request, header, headers, body, bodyReader
-    , jsonData
+    , jsonData, formData
       -- ** Accessing Path, Form and Query Parameters
     , param, params
     , pathParam, captureParam, formParam, queryParam
     , pathParamMaybe, captureParamMaybe, formParamMaybe, queryParamMaybe
     , pathParams, captureParams, formParams, queryParams
       -- *** Files
-    , files, filesOpts, Trans.ParseRequestBodyOptions
+    , files, filesOpts
       -- ** Modifying the Response and Redirecting
     , status, addHeader, setHeader, redirect
       -- ** Setting Response Body
@@ -52,6 +52,8 @@ module Web.Scotty
       -- * Types
     , ScottyM, ActionM, RoutePattern, File, Content(..), Kilobytes, ErrorHandler, Handler(..)
     , ScottyState, defaultScottyState
+    -- ** Functions from Cookie module
+    , setSimpleCookie,getCookie,getCookies,deleteCookie,makeSimpleCookie
     ) where
 
 import qualified Web.Scotty.Trans as Trans
@@ -67,10 +69,12 @@ import Network.HTTP.Types (Status, StdMethod, ResponseHeaders)
 import Network.Socket (Socket)
 import Network.Wai (Application, Middleware, Request, StreamingBody)
 import Network.Wai.Handler.Warp (Port)
-import qualified Network.Wai.Parse as W (defaultParseRequestBodyOptions)
+import qualified Network.Wai.Parse as W
 
+import Web.FormUrlEncoded (FromForm)
 import Web.Scotty.Internal.Types (ScottyT, ActionT, ErrorHandler, Param, RoutePattern, Options, defaultOptions, File, Kilobytes, ScottyState, defaultScottyState, ScottyException, StatusError(..), Content(..))
 import UnliftIO.Exception (Handler(..), catch)
+import Web.Scotty.Cookie (setSimpleCookie,getCookie,getCookies,deleteCookie,makeSimpleCookie)
 
 {- $setup
 >>> :{
@@ -242,7 +246,7 @@ files = Trans.files
 -- | Get list of temp files and form parameters decoded from multipart payloads.
 --
 -- NB the temp files are deleted when the continuation exits
-filesOpts :: Trans.ParseRequestBodyOptions
+filesOpts :: W.ParseRequestBodyOptions
           -> ([Param] -> [File FilePath] -> ActionM a) -- ^ temp files validation, storage etc
           -> ActionM a
 filesOpts = Trans.filesOpts
@@ -272,6 +276,12 @@ bodyReader = Trans.bodyReader
 -- NB: uses 'body' internally
 jsonData :: FromJSON a => ActionM a
 jsonData = Trans.jsonData
+
+-- | Parse the request body as @x-www-form-urlencoded@ form data and return it. Raises an exception if parse is unsuccessful.
+--
+-- NB: uses 'body' internally
+formData :: FromForm a => ActionM a
+formData = Trans.formData
 
 -- | Get a parameter. First looks in captures, then form data, then query parameters.
 --